Backend/DocumentService/requirements.txt
flask
flask_sqlalchemy
werkzeug
requests
psycopg2-binary
python-pptx
openai